requirements.txt
streamlit
pandas
numpy
plotly
fredapi
python-dotenv